requirements.txt
streamlit
streamlit-agraph
numpy
pandas
faiss-cpu
nltk
sentence_transformers
